fix: bugs of deleting events
