+++ src/store/modalStore.ts
@@ -1,19 +1,5 @@
 import { create } from 'zustand';
-import { Event } from '../utils/type';
-
-export const initialEvent = {
-  eventId: 0,
-  title: '',
-  startAt: new Date(),
-  endAt: new Date(),
-  isAllDay: false,
-  isMemo: false,
-  tag: '0',
-  note: '',
-  createdAt: null,
-  updatedAt: null,
-  messages: [],
-};
+import { Event, initialEvent } from '../utils/type';
 
 interface ModalState {
   isCreateModalOpen: boolean;
